pyproject.toml
[project]
name = "nonebot_plugin_smart_reply"
version = "0.15.114514"
description = "nonebot2的融合了openai, newbing, 词库的智障回复插件"
authors = [
    {name = "Special-Week", email = "HuaMing27499@gmail.com"},
]
dependencies = [
    "revChatGPT>=3.3.5",
    "pillow>=9.1.1",
    "nonebot2>=2.0.0b5",
    "nonebot-adapter-onebot>=2.2.3",
    "httpx"
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}
